vst3: fix some incorrect refcount logic.
if we are passed some object as an argument, we are borrowing it,
hence if we want to hold onto it after returning we must increment
its refcount. on the other hand, if we receive some object as a
return value, we must adopt it, meaning we are responsible for
eventually decrementing its refcount.