maintain preallocated Vecs of AudioBus and AudioBuffer
In order to avoid allocations in process(), we need to keep around large
enough buffers to store AudioBuses and AudioBuffers in. We ensure these
buffers are the right size at initialization and when switching bus
layouts.

The AudioBuses' and AudioBuffers' lifetime parameters need to be static
at rest, but they need to be shorter for the duration of process(), so
that the AudioBuses can hold references to the AudioBuffers. To
accomplish this, we transmute the vectors to have the necessary lifetime
parameters before setting them up, and then we clear them and transmute
them back afterwards.