restructure audio buffers again
Use the slices of pointers passed to us from the host rather than
storing our own array of pointers. Simplifies things a lot and makes it
so that we no longer have to resize a buffer when changing bus layouts
or transmute Vec references when building the AudioBuffers structure in
process().