clap: create a static for Factory and create/destroy it in init/deinit.
The clap_plugin_descriptor needs to be created at runtime based on
Plugin::info(), and the only place CLAP gives us to clean it up is
the clap_plugin_factory's deinit. Since init/deinit take no arguments,
the relevant state needs to be stored in a static, which in turn needs
to be declared in the clap! entry point macro.