redesign params API
Under the previous design, a set of parameter objects would be stored in
the Plugin object, and then the ParamList would hold a set of trait
objects for accessing these parameter objects. Under the new design,
only the current values of the parameters will be stored in the Plugin,
and all of the parameter metadata and behavior will be stored in
ParamInfo structs in the ParamList.

ParamInfo holds metadata like the id, name, and label, as well as a set
of trait objects for defining a parameter's mapping between normalized
and plain, how to convert it to or from text, and how to set or get it
on the Plugin object.

The ParamList is now stored in the PluginHandle and accessible via
PluginHandle::params();