switch from supports_layout() to bus_configs()
Instead of having a method that is passed the candidate bus
configuration and returns a bool for whether it is supported, have the
plugin construct a static list of supported bus configurations and then
check subsequent candidate configurations against this list.

This makes it easier to support CLAP's audio port configuration system,
as well as some aspects of AudioUnit's and AAX's bus configuration
systems. Eventually it might be desirable to make this system a bit more
flexible, to allow for an open set of configurations chosen
programmatically at runtime by either the host or the plugin, but this
will work for now.