vst3: always deliver parameter edits to Plugin
Don't bother with only delivering parameter edits to the Plugin if the
audio processor is not active. Since the Processor gets sample-accurate
automation events and holds its own copy of the parameter values, the
precise timing of when parameter edits are delivered to the Plugin does
not matter for automation playback. The Plugin should hold the most
up-to-date copy of the parameter values, and it doesn't matter if values
sometimes get set twice from both the audio thread and main thread.