API change: make Plugin a single-threaded object
Previously, the Plugin object had to be Sync, and the set_param and load
methods took &self, meaning that Plugin had to be implemented in a
thread-safe way, using atomics, etc. for its internal state. After this
change, Plugin is now a single-threaded object, and set_param and load
take &mut self. It is still required to be Send, as save and load won't
necessarily always happen on the main thread in all APIs, but it no
longer has to be thread-safe.

The methods for constructing the Editor and Processor have been moved to
the Plugin trait, to emphasize that they take place in the same
single-threaded context as the rest of Plugin's methods.