new bus api: combine input and output buses into a single flat list
Switch from having a separate list of input buses and output buses to having a
single flat list of buses, each specifying whether it is an input or an output.
This will make it possible to add an InOut bus type representing a matching
pair of input and output buses, having the same channel count, which can be
processed in-place.