support in-out buses
Add support for an InOut bus direction. In-out buses represent a matched pair
of input and output buses, always having the same channel count, which are
presented to the plugin as a single set of mutable buffers for in-place
processing. When the host provides a single set of buffers for in-place
processing, these can be passed straight through to the plugin with zero buffer
copies; when the post passes separate input and output buffers, the inputs will
be copied to the outputs before being passed to the plugin.