handle aliased input buffers when copying inputs to outputs
For in-out buses, if a pair of input and output pointers are not equal, we have
to copy the contents of the input buffer to its corresponding output buffer. If
an input buffer is aliased by a different output buffer than the one it
corresponds to (which can happen if the host and plugin disagree on the mapping
between input and output channels), we might end up overwriting it before it
gets copied.

To prevent this, we check each input pointer supplied to an in-out bus against
the set of output pointers, and if it's aliased, we copy it to a scratch buffer
before performing any copies.