vst3: separate copy of param values for IEditController
Instead of calling Plugin::set_param and get_param in
IEditController::setParamNormalized/getParamNormalized, maintain a separate set
of param values for the IEditController (which will also be used as the current
param values for the Editor). Since we are still a single-component effect,
continue to do nothing in IEditController::setComponentState, and update the
IEditController param values to match the Plugin's param values in
IComponent::setState. This matches the behavior of the SingleComponentEffect
class in the VST 3 SDK.

After this change, parameter state is no longer being synchronized between the
Processor and Plugin, so edits and automation are not currently applied to the
Plugin, and patch loads that occur while the Processsor exists will not be
applied to the Processor. This will be addressed in a future change.