vst3: parameter synchronization between Plugin and Processor
Define a ParamValues struct for bidirectional parameter state synchronization
between the Plugin and Processor structs. Current parameter values are held in
a set of AtomicF64s, and each struct has a corresponding AtomicBitset of dirty
flags to efficiently poll for changes to parameter values.

On the Processor side, apply parameter change events to ParamValues in
IAudioProcessor::process, and poll for any changes from the Plugin before
calling Plugin::process or reset. On the Plugin side, write out the set of
parameter values to ParamValues after a successful call to Plugin::load, and
poll for any changes from the Processor before calling any method which could
be affected by the current set of parameter values (e.g. Plugin::save,
processor, editor, and latency).

Eventually, ParamValues::sync_plugin should be called regularly from a timer on
the main thread, so that the Plugin's state is kept up to date; however, this
shouldn't be necessary for correctness, as sync_plugin is also called before
any Plugin method that requires up-to-date parameter values.