clap: improvements regarding parameter synchronization and activate/deactivate
Remove unnecessary calls to `sync_plugin` in `activate` and in
main-thread calls to `flush`. These methods can only be called while the
plugin is inactive, so there should not be any parameter changes to poll
for.

Discard any enqueued plugin->processor changes in `activate`, since
these changes should already be reflected in the initial state of the
processor, and call `sync_plugin` in `deactivate` to ensure that all
processor->plugin parameter changes have been applied before entering
the inactive state.